requirements.txt
google-api-python-client
langchain
python-dotenv
googlemaps
fastapi
pydantic
uvicorn
openai
beautifulsoup4
twilio